+++ gamemodes/guesswho/entities/weapons/weapon_gw_shrink.lua
@@ -4,8 +4,8 @@ SWEP.AbilitySound = "gwabilities/smb2_shrink.wav"
 
 function SWEP:Ability()
     timer.Create( "Ability.Effect." .. self.Owner:SteamID(), 7, 1, function() self:OnRemove() end )
-    self.Owner:SetRunSpeed( self.Owner:GetRunSpeed() * 1.5 )
-    self.Owner:SetWalkSpeed( self.Owner:GetWalkSpeed() * 1.5 )
+    self.Owner:SetRunSpeed( self.Owner:GetRunSpeed() * 1.75 )
+    self.Owner:SetWalkSpeed( self.Owner:GetWalkSpeed() * 1.75 )
     self.Owner:SetModelScale( self.Owner:GetModelScale() / 2.5, 1 )
     self.Owner:SetHealth( self.Owner:Health() / 2.5 )
 end